Removed warnings

+++ HW02/convolution.cpp
@@ -3,14 +3,14 @@
 
 using namespace std;
 void convolve(const float *image, float *output, std::size_t n, const float *mask, std::size_t m){
-    int m2= (m)/2;
-    for(int i=0; i<n; i++){
-        for(int j=0; j<n; j++){
+    size_t m2= (m)/2;
+    for(size_t i=0; i<n; i++){
+        for(size_t j=0; j<n; j++){
             float sum=0;
-            for(int k=0; k<m; k++){
-                for(int l=0; l<m; l++){
-                    int r= i+k-m2;
-                    int c= j+l-m2;
+            for(size_t k=0; k<m; k++){
+                for(size_t l=0; l<m; l++){
+                    size_t r= i+k-m2;
+                    size_t c= j+l-m2;
                     if((r<0 || r>=n) && (c<0 || c>=n)){
                         sum+=mask[k*m+l]*0;
                     }
